refactor(3.10): clean up
